code refactoring, IntellijPluginManager class
