+++ cli/src/main/java/com/devonfw/tools/ide/tool/intellij/Intellij.java
@@ -1,8 +1,6 @@
 package com.devonfw.tools.ide.tool.intellij;
 
 import java.io.IOException;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Set;
@@ -29,6 +27,8 @@ public class Intellij extends IdeToolCommandlet {
   private static final String IDEA_BASH_SCRIPT = IDEA + ".sh";
   private static final String BUILD_FILE = "build.txt";
 
+  private final IntellijPluginManager pluginManager;
+
   /**
    * The constructor.
    *
@@ -37,6 +37,7 @@ public class Intellij extends IdeToolCommandlet {
   public Intellij(IdeContext context) {
 
     super(context, "intellij", Set.of(Tag.INTELLIJ));
+    this.pluginManager = new IntellijPluginManager(context, this);
   }
 
   @Override
@@ -89,7 +90,7 @@ public class Intellij extends IdeToolCommandlet {
     }
   }
 
-  private String generateMacEditionString() {
+  String generateMacEditionString() {
 
     String edition = "";
     if (getConfiguredEdition().equals("intellij")) {
@@ -112,106 +113,7 @@ public class Intellij extends IdeToolCommandlet {
 
   @Override
   public void installPlugin(PluginDescriptor plugin) {
-
-    String pluginId = plugin.getId();
-    String downloadUrl = plugin.getUrl();
-
-    Path tmpDir = null;
-    try {
-      String buildVersion = readBuildVersion();
-
-      Path installationPath = getPluginsInstallationPath();
-      ensureInstallationPathExists(installationPath);
-
-      if (downloadUrl == null || downloadUrl.isEmpty()) {
-        downloadUrl = String.format("https://plugins.jetbrains.com/pluginManager?action=download&id=%s&build=%s", pluginId, buildVersion);
-      }
-
-      FileAccess fileAccess = context.getFileAccess();
-      tmpDir = fileAccess.createTempDir(pluginId);
-
-      Path downloadedFile = downloadPlugin(fileAccess, downloadUrl, tmpDir, buildVersion, pluginId);
-      installDownloadedPlugin(fileAccess, downloadedFile, pluginId);
-
-
-    } catch (IOException e) {
-      throw new IllegalStateException("Failed to process plugin installation", e);
-    } finally {
-      if (tmpDir != null) {
-        context.getFileAccess().delete(tmpDir);
-      }
-    }
-  }
-
-  private String readBuildVersion() throws IOException {
-
-    Path buildFile = getToolPath().resolve(BUILD_FILE);
-    if (this.context.getSystemInfo().isMac()) {
-      buildFile = getToolPath().resolve("IntelliJ IDEA" + generateMacEditionString() + ".app").resolve("Contents").resolve("Resources").resolve(BUILD_FILE);
-    }
-    try {
-      return Files.readString(buildFile);
-    } catch (IOException e) {
-      throw new IllegalStateException("Failed to read IntelliJ build version: " + buildFile, e);
-    }
-  }
-
-  private void ensureInstallationPathExists(Path installationPath) throws IOException {
-    if (!Files.exists(installationPath)) {
-      try {
-        Files.createDirectories(installationPath);
-      } catch (IOException e) {
-        throw new IllegalStateException("Failed to create directory " + installationPath, e);
-      }
-    }
-  }
-
-  private Path downloadPlugin(FileAccess fileAccess, String downloadUrl, Path tmpDir, String buildVersion, String pluginId) throws IOException {
-    String extension = getFileExtensionFromUrl(downloadUrl);
-    if (extension.isEmpty()) {
-      throw new IllegalStateException("Unknown file type for URL: " + downloadUrl);
-    }
-    String fileName = String.format("intellij-plugin-%s-%s%s", buildVersion, pluginId, extension);
-    Path downloadedFile = tmpDir.resolve(fileName);
-    fileAccess.download(downloadUrl, downloadedFile);
-    return downloadedFile;
-  }
-
-  private void installDownloadedPlugin(FileAccess fileAccess, Path downloadedFile, String pluginId) throws IOException {
-    Path targetDir = getPluginsInstallationPath().resolve(pluginId);
-    if (Files.exists(targetDir)) {
-      context.info("File already exists");
-    } else {
-      fileAccess.extract(downloadedFile, targetDir);
-    }
-  }
-
-  private String getFileExtensionFromUrl(String urlString) throws IOException {
-
-    URL url = new URL(urlString);
-    HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-    connection.setRequestMethod("HEAD");
-    connection.connect();
-
-    int responseCode = connection.getResponseCode();
-    if (responseCode != HttpURLConnection.HTTP_OK) {
-      throw new IOException("Failed to fetch file headers: HTTP " + responseCode);
-    }
-
-    String contentType = connection.getContentType();
-    return getFileExtensionFromContentType(contentType);
-  }
-
-  private String getFileExtensionFromContentType(String contentType) {
-
-    if (contentType == null) {
-      return "";
-    }
-    return switch (contentType) {
-      case "application/zip" -> ".zip";
-      case "application/java-archive" -> ".jar";
-      default -> "";
-    };
+    pluginManager.installPlugin(plugin);
   }
 
 }
